Use shared projects to properly share the code between the two assemblies
