+++ src/LaunchDarkly.Common.Shared/EventUser.cs
