+++ src/LaunchDarkly.Common.Shared/IBaseConfiguration.cs
