+++ src/LaunchDarkly.Common.Shared/UnsuccessfulResponseException.cs
