+++ src/LaunchDarkly.Common.Shared/Util.cs
