+++ src/LaunchDarkly.Common.StrongName/LaunchDarkly.Common.StrongName.csproj
@@ -11,14 +11,12 @@
     <Copyright>Copyright 2018 LaunchDarkly</Copyright>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="..\LaunchDarkly.Common\*.cs" Exclude="..\LaunchDarkly.Common\AssemblyInfo.cs" />
-  </ItemGroup>
-  <ItemGroup>
     <PackageReference Include="Common.Logging" Version="3.4.1" />
     <PackageReference Include="LaunchDarkly.EventSource" Version="3.1.5" />
     <PackageReference Include="Newtonsoft.Json" Version="9.0.1" Condition="'$(TargetFramework)' != 'net45'" />
     <PackageReference Include="Newtonsoft.Json" Version="6.0.1" Condition="'$(TargetFramework)' == 'net45'" />
   </ItemGroup>
+  <Import Project="..\LaunchDarkly.Common.Shared\LaunchDarkly.Common.Shared.projitems" Label="Shared" />
   <ItemGroup Condition=" '$(TargetFramework)' == 'net45' ">
     <Reference Include="System" />
     <Reference Include="Microsoft.CSharp" />
